Add ChatPluginController
